fix tests on linux

+++ src/file_lock.rs
@@ -56,7 +56,7 @@ pub(crate) fn getlk(fd: RawFd, requested: FileLock) -> Result<FileLock, libc::c_
         let length = u64::try_from(lock.l_len).map_err(|_| libc::EIO)?;
         reply_start.checked_add(length - 1).ok_or(libc::EOVERFLOW)?
     };
-    let kind = match lock.l_type {
+    let kind = match lock.l_type as libc::c_int {
         libc::F_RDLCK => LockKind::Read,
         libc::F_WRLCK => LockKind::Write,
         libc::F_UNLCK => LockKind::Unlock,
